numerous tweaks and improvements along with Debian 12-only support
The larger issue is related to the reverse proxy dropping the audio.
When running without a proxy the audio continues but the "Tune In" button reappears.

I disabled the timeout from 0ms and added some additional logging along with a backup method to play audio (click the logo).
The systemd issue on installation is resolved and the 'python3-jinja' change was implemented.