attempt to solve the audio failover behaviour when reloading
Earlier this year I attempted to implement some default radio
stations to use when the locally generated audio was still un-
populated. Rather than tell the user that they need to wait 8
hours to see results it made sense to default to the local audio
but have some fallback audio sources that can be expected to be
available.

It seems when the audio cuts out it loops through the audio sources
again when all it needs to do is just keep playing. I think I have
solved this by simplifying the loop among other tweaks. Observed as
stable on a test station.