hotfix: add temporary modification to install script
Unannounced change to BBC URLs forced an immediate update to the URLs
used. The servers I had running immediately failed and this commit
has been tested on a separate beta service with satisfactory results.

v1.3.2 will likely include the permanent URL change as a new hard-coded
value and a fix for the "local audio not buffered" on my primary
servers.